+++ age/src/primitives/stream.rs
@@ -565,13 +565,22 @@ impl<R: Read + Seek> StreamReader<R> {
                 let num_chunks =
                     (ct_len + (ENCRYPTED_CHUNK_SIZE as u64 - 1)) / ENCRYPTED_CHUNK_SIZE as u64;
 
+                // If we have no ciphertext data then there is no last chunk, which is
+                // invalid.
+                let non_last_chunks = num_chunks.checked_sub(1).ok_or_else(|| {
+                    io::Error::new(
+                        io::ErrorKind::InvalidData,
+                        "Last chunk is invalid, stream might be truncated",
+                    )
+                })?;
+
                 // Authenticate the ciphertext length by checking that we can successfully
                 // decrypt the last chunk _as_ a last chunk.
-                let last_chunk_start = ct_start + ((num_chunks - 1) * ENCRYPTED_CHUNK_SIZE as u64);
+                let last_chunk_start = ct_start + (non_last_chunks * ENCRYPTED_CHUNK_SIZE as u64);
                 let mut last_chunk = Vec::with_capacity((ct_end - last_chunk_start) as usize);
                 self.inner.seek(SeekFrom::Start(last_chunk_start))?;
                 self.inner.read_to_end(&mut last_chunk)?;
-                self.stream.nonce.set_counter(num_chunks - 1);
+                self.stream.nonce.set_counter(non_last_chunks);
                 self.stream.decrypt_chunk(&last_chunk, true).map_err(|_| {
                     io::Error::new(
                         io::ErrorKind::InvalidData,
@@ -679,7 +688,7 @@ mod tests {
     use age_core::secrecy::ExposeSecret;
     use std::io::{self, Cursor, Read, Seek, SeekFrom, Write};
 
-    use super::{PayloadKey, Stream, CHUNK_SIZE};
+    use super::{PayloadKey, Stream, CHUNK_SIZE, TAG_SIZE};
 
     #[cfg(feature = "async")]
     use futures::{
@@ -1017,6 +1026,44 @@ mod tests {
             }
             Ok(_) => panic!("This is a security issue."),
         }
+    }
+
+    #[test]
+    fn seek_from_end_with_empty_fails_on_truncation() {
+        // The empty plaintext means we should encrypt to a single chunk.
+        let plaintext: Vec<u8> = b"".to_vec();
+
+        // Encrypt the plaintext just like the example code in the docs.
+        let mut encrypted = vec![];
+        {
+            let mut w = Stream::encrypt(PayloadKey([7; 32].into()), &mut encrypted);
+            w.write_all(&plaintext).unwrap();
+            w.finish().unwrap();
+        };
+        assert_eq!(encrypted.len(), TAG_SIZE);
+
+        // Every truncated length should result in an error.
+        for i in 0..TAG_SIZE {
+            let truncated_ciphertext = &encrypted[..i];
+            let mut truncated_reader = Stream::decrypt(
+                PayloadKey([7; 32].into()),
+                Cursor::new(truncated_ciphertext),
+            );
+            match truncated_reader.seek(SeekFrom::End(0)) {
+                Err(e) => {
+                    assert_eq!(e.kind(), io::ErrorKind::InvalidData);
+                    assert_eq!(
+                        &e.to_string(),
+                        "Last chunk is invalid, stream might be truncated",
+                    );
+                }
+                Ok(_) => panic!("This is a security issue."),
+            }
+        }
+
+        // Decrypting without truncation should show an empty file.
+        let mut reader = Stream::decrypt(PayloadKey([7; 32].into()), Cursor::new(encrypted));
+        assert_eq!(reader.len().unwrap(), 0);
     }
 
     #[test]